services/worker/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.27
alembic==1.13.1
python-dotenv==1.0.1
telethon==1.32.1
aiohttp==3.9.3
pydantic==2.6.1 